+++ projects/angular8-yandex-maps/src/lib/components/yandex-control-component/yandex-control.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
+import { ILoadEvent } from '../../types/types';
 
 @Component({
   selector: 'angular-yandex-control',
@@ -9,7 +10,7 @@ export class YandexControlComponent implements OnInit {
   @Input() public type: string;
   @Input() public parameters: any;
 
-  @Output() public load = new EventEmitter<any>();
+  @Output() public load = new EventEmitter<ILoadEvent>();
 
   constructor() {}
   public ngOnInit(): void {}
@@ -17,7 +18,12 @@ export class YandexControlComponent implements OnInit {
   public initControl(ymaps: any, map: any): void {
     const control = new ymaps.control[this.type](this.parameters);
 
+    // RoutePanel ignores state in parameters. Probably API bug
+    if (this.type === 'RoutePanel' && this.parameters && this.parameters.state) {
+      control.routePanel.state.set({ ...this.parameters.state });
+    }
+
     map.controls.add(control);
-    this.load.emit(control);
+    this.load.emit({ ymaps, instance: control });
   }
 }