+++ projects/angular8-yandex-maps/src/lib/components/yandex-multiroute-component/yandex-multiroute.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnInit, Output, EventEmitter } from '@angular/core';
-import { IEvent } from '../../types/types';
+import { IEvent, ILoadEvent } from '../../types/types';
 
 @Component({
   selector: 'angular-yandex-multiroute',
@@ -11,7 +11,7 @@ export class YandexMultirouteComponent implements OnInit {
   @Input() public model: any;
   @Input() public options: any;
 
-  @Output() public load = new EventEmitter<any>();
+  @Output() public load = new EventEmitter<ILoadEvent>();
   @Output() public activeroutechange = new EventEmitter<IEvent>();
   @Output() public baloon = new EventEmitter<IEvent>();
   @Output() public yaclick = new EventEmitter<IEvent>();
@@ -37,49 +37,50 @@ export class YandexMultirouteComponent implements OnInit {
     );
 
     map.geoObjects.add(multiroute);
-    this.emitEvents(multiroute);
+    this.emitEvents(ymaps, multiroute);
   }
 
   /**
    * Emit events
+   * @param ymaps - class from Yandex.Map API
    * @param multiroute - multiroute instance
    */
-  public emitEvents(multiroute: any): void {
-    this.load.emit(multiroute);
+  public emitEvents(ymaps: any, multiroute: any): void {
+    this.load.emit({ ymaps, instance: multiroute });
 
     // Activeroutechange
     multiroute.events
       .add(
         'activeroutechange',
-        (e: any) => this.activeroutechange.emit({ instance: multiroute, type: e.originalEvent.type, event: e })
+        (e: any) => this.activeroutechange.emit({ ymaps, instance: multiroute, type: e.originalEvent.type, event: e })
       );
 
     // Baloon
     multiroute.events
       .add(
         ['balloonopen', 'balloonclose'],
-        (e: any) => this.baloon.emit({ instance: multiroute, type: e.originalEvent.type, event: e })
+        (e: any) => this.baloon.emit({ ymaps, instance: multiroute, type: e.originalEvent.type, event: e })
       );
 
     // Click
     multiroute.events
       .add(
         ['click', 'dblclick'],
-        (e: any) => this.yaclick.emit({ instance: multiroute, type: e.originalEvent.type, event: e })
+        (e: any) => this.yaclick.emit({ ymaps, instance: multiroute, type: e.originalEvent.type, event: e })
       );
 
     // Mouse
     multiroute.events
       .add(
         ['mousedown', 'mouseenter', 'mouseleave', 'mousemove', 'mouseup'],
-        (e: any) => this.mouse.emit({ instance: multiroute, type: e.originalEvent.type, event: e })
+        (e: any) => this.mouse.emit({ ymaps, instance: multiroute, type: e.originalEvent.type, event: e })
       );
 
     // Multitouch
     multiroute.events
       .add(
         ['multitouchstart', 'multitouchmove', 'multitouchend'],
-        (e: any) => this.multitouch.emit({ instance: multiroute, type: e.originalEvent.type, event: e })
+        (e: any) => this.multitouch.emit({ ymaps, instance: multiroute, type: e.originalEvent.type, event: e })
       );
   }
 }