+++ projects/angular8-yandex-maps/src/lib/components/yandex-placemark-component/yandex-placemark.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
-import { IEvent } from '../../types/types';
+import { IEvent, ILoadEvent } from '../../types/types';
 
 @Component({
   selector: 'angular-yandex-placemark',
@@ -11,7 +11,7 @@ export class YandexPlacemarkComponent implements OnInit {
   @Input() public properties: any;
   @Input() public options: any;
 
-  @Output() public load = new EventEmitter<any>();
+  @Output() public load = new EventEmitter<ILoadEvent>();
   @Output() public baloon = new EventEmitter<IEvent>();
   @Output() public yaclick = new EventEmitter<IEvent>();
   @Output() public drag = new EventEmitter<IEvent>();
@@ -36,58 +36,59 @@ export class YandexPlacemarkComponent implements OnInit {
     const placemark = new ymaps.Placemark(this.geometry, this.properties, this.options);
 
     map.geoObjects.add(placemark);
-    this.emitEvents(placemark);
+    this.emitEvents(ymaps, placemark);
 
     return placemark;
   }
 
   /**
    * Emit events
+   * @param ymaps - class from Yandex.Map API
    * @param placemark - placemark instance
    */
-  public emitEvents(placemark: any): void {
-    this.load.emit(placemark);
+  public emitEvents(ymaps: any, placemark: any): void {
+    this.load.emit({ ymaps, instance: placemark });
 
     // Baloon
     placemark.events
       .add(
         ['balloonopen', 'balloonclose'],
-        (e: any) => this.baloon.emit({ instance: placemark, type: e.originalEvent.type, event: e })
+        (e: any) => this.baloon.emit({ ymaps, instance: placemark, type: e.originalEvent.type, event: e })
       );
 
     // Click
     placemark.events
       .add(
         ['click', 'dblclick'],
-        (e: any) => this.yaclick.emit({ instance: placemark, type: e.originalEvent.type, event: e })
+        (e: any) => this.yaclick.emit({ ymaps, instance: placemark, type: e.originalEvent.type, event: e })
       );
 
     // Drag
     placemark.events
       .add(
         ['dragstart', 'dragend'],
-        (e: any) => this.drag.emit({ instance: placemark, type: e.originalEvent.type, event: e })
+        (e: any) => this.drag.emit({ ymaps, instance: placemark, type: e.originalEvent.type, event: e })
       );
 
     // Hint
     placemark.events
       .add(
         ['hintopen', 'hintclose'],
-        (e: any) => this.hint.emit({ instance: placemark, type: e.originalEvent.type, event: e })
+        (e: any) => this.hint.emit({ ymaps, instance: placemark, type: e.originalEvent.type, event: e })
       );
 
     // Mouse
     placemark.events
       .add(
         ['mousedown', 'mouseenter', 'mouseleave', 'mousemove', 'mouseup'],
-        (e: any) => this.mouse.emit({ instance: placemark, type: e.originalEvent.type, event: e })
+        (e: any) => this.mouse.emit({ ymaps, instance: placemark, type: e.originalEvent.type, event: e })
       );
 
     // Multitouch
     placemark.events
       .add(
         ['multitouchstart', 'multitouchmove', 'multitouchend'],
-        (e: any) => this.multitouch.emit({ instance: placemark, type: e.originalEvent.type, event: e })
+        (e: any) => this.multitouch.emit({ ymaps, instance: placemark, type: e.originalEvent.type, event: e })
       );
   }
 }